requirements.txt
celery[redis]
redis
aiohttp
beautifulsoup4
python-dotenv
fastapi
uvicorn[standard]
flower
curl-cffi
httpx